% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AudioScatterV6.R
\docType{data}
\name{AudioData}
\alias{AudioData}
\title{AudioData}
\format{
A data frame with rows and 4 columns
}
\usage{
data(AudioData)
}
\description{
An example dataset containing the preop PTA, preop WRS, postop PTA, and postop WRS.
You can load your own dataset containing the preop PTA, preop WRS, and potentially postop PTA and postop WRS.
The variables are as follows:
}
\details{
\itemize{
  \item PrePTA. preop pure tone average
  \item PreWRS. preop word recognition score
  \item PostPTA. postop pure tone average
  \item PostWRS. postop word recognition score
}
}
\keyword{datasets}
