% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AudioScatterV5.R
\name{CalcChange}
\alias{CalcChange}
\title{Post-treatment change in PTA and WRS
Calculate change in PTA
Calculate change in WRS
create new columns relating the change in WRS to the correct column for the scattergram.
create new columns relating the change in PTA to the correct row for the scattergram.
Calculate Change between Post and Pre PTA and WRS and put in appropriate column and row for scattergram.}
\usage{
CalcChange(dTable)
}
\arguments{
\item{dTable}{table loaded from csv file including PrePTA, PreWRS, and possibly PostPTA, PostWRS.}
}
\description{
Post-treatment change in PTA and WRS
Calculate change in PTA
Calculate change in WRS
create new columns relating the change in WRS to the correct column for the scattergram.
create new columns relating the change in PTA to the correct row for the scattergram.
Calculate Change between Post and Pre PTA and WRS and put in appropriate column and row for scattergram.
}
