% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AudioScatterV5.R
\docType{data}
\name{dTable}
\alias{dTable}
\title{# load example dataset to dTable.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 58 rows and 4 columns.
}
\usage{
dTable
}
\description{
# load example dataset to dTable.
}
\keyword{datasets}
