% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AudioScatterV5.R
\docType{data}
\name{dTablePost}
\alias{dTablePost}
\title{Calculate the audiogram changes and write to new table dTablePost}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 58 rows and 8 columns.
}
\usage{
dTablePost
}
\description{
Calculate the audiogram changes and write to new table dTablePost
}
\keyword{datasets}
