% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AudioScatterV5.R
\name{SavePostScatter}
\alias{SavePostScatter}
\title{SavePostScatter}
\usage{
SavePostScatter(dTablePost)
}
\arguments{
\item{dTablePost}{table resulting from CalcChange with change in PTA and WRS, dPTA and dWRS, respectively, as well as the column and row dPTA and dWRS should go.}
}
\value{
png file of postop scattergram
}
\description{
SavePostScatter
}
