% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AudioScatterV6.R
\name{SavePreScatter}
\alias{SavePreScatter}
\title{SavePreScatter}
\usage{
SavePreScatter(dTable)
}
\arguments{
\item{dTable}{table loaded from csv file including PrePTA, PreWRS, and possibly PostPTA, PostWRS.}
}
\value{
png file for preop scattergram.
}
\description{
SavePreScatter
}
